[App] [AssemblyPart] Fixed exiting of loop when wrong H is determined. Added IoU calculation and distance calculation for oclusion demonstration [App] [AugmentedInstructions] Added written instructions. Fixed Instructions for with CalculateCenter and so on (problem with idx in the loops), works for all intended sides. Added videos to play at each step [App] [ALL] Static Casting for some values

+++ OpenCV_ZEDSDK/AugmentedAssembly/include/AugmentedAssembly.h
@@ -131,7 +131,7 @@ private:
 
         if(input.getPtr<sl::uchar1>(sl::MEM::CPU) != NULL)
         {
-            return cv::Mat(input.getHeight(), input.getWidth(), cv_type, input.getPtr<sl::uchar1>(sl::MEM::CPU));
+            return cv::Mat(static_cast<int>(input.getHeight()), static_cast<int>(input.getWidth()), cv_type, input.getPtr<sl::uchar1>(sl::MEM::CPU));
         }
         return ret;
     }
